+++ docs/attribute_curator/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this gear are documented in this file.
     * Adds `img_study_date` to use the `file.info.header.dicom.StudyDate` field for image file ordering
     * Adds `file.info._filename` as temporary metadata added to the table for deriver use (needed for curation of images file locators)
 * Queries RxClass API for V4+ A4 derived variable calculations, and passes to the deriver through the table under `_rxclass`
+    * Adds optional input JSON `rxclass_concepts` to provide cached results to skip querying, since this step takes a while otherwise
 * Adds loop for missingness logic
     * Passes the previous record to the deriver to support missingness needs through the table under `_prev_record`
 * Updates function names from `pre/post_process` to `pre/post_curate`